+++ packages/flopper/src/presetFlopperMaterial.js
@@ -1,32 +1,32 @@
-import { flop, rand }                          from '@aryth/rand'
 import { NUM_DESC }                            from '@aryth/comparer'
-import { Grey }       from '@palett/cards'
-import { HEX }        from '@palett/enum-color-space'
-import { randPreset } from '@palett/presets'
+import { flop, rand }                          from '@aryth/rand'
+import { Grey }                                from '@palett/cards'
+import { HEX }                                 from '@palett/enum-color-space'
+import { randPres }                            from '@palett/pres'
 import { ColorGroups, Degrees, palettCrostab } from '@palett/table'
-import { swap }           from '@vect/swap'
-import { degreeToIndice } from './utils/degreeToIndice.js'
-import { sortBy }         from './utils/sortDegrees.js'
+import { swap }                                from '@vect/swap'
+import { degreeToIndice }                      from './utils/degreeToIndice.js'
+import { sortBy }                              from './utils/sortDegrees.js'
 
 export function* presetFlopperMaterial({
-                                 degrees = Degrees.entire,
-                                 colors = ColorGroups.rainbow,
-                                 space = HEX,
-                                 defaultColor = Grey.lighten_1,
-                                 exhausted = true
-                               } = {}) {
-  const crostab = palettCrostab({space,  degrees, colors, dyed: false })
+                                         degrees = Degrees.entire,
+                                         colors = ColorGroups.rainbow,
+                                         space = HEX,
+                                         defaultColor = Grey.lighten_1,
+                                         exhausted = true
+                                       } = {}) {
+  const crostab = palettCrostab({ space, degrees, colors, dyed: false })
   degrees = sortBy.call(degrees.slice(), degreeToIndice, NUM_DESC)
   let h = degrees.length, w = colors.length
   for (let i = 0; i < h; i++) {
     for (let j = w - 1, side = degrees[i], head = crostab.head.slice(); j >= 0; j--) {
       const banner = swap.call(head, rand(j), j)
       const hex = crostab.cell(side, banner)
-      yield randPreset(hex)
+      yield randPres(hex)
     }
   }
   defaultColor = defaultColor ?? (crostab.cell(degrees[0], flop(crostab.head)))
-  const defaultPreset = randPreset(defaultColor)
+  const defaultPreset = randPres(defaultColor)
   while (!exhausted) yield defaultPreset
   return defaultPreset
 }