+++ packages/flopper/test/shiftFlopper.test.js
@@ -1,9 +1,9 @@
+import { init }         from '@vect/vector-init'
 import { test }         from 'node:test'
 import { shiftFlopper } from '../src/utils/shiftFlopper.js'
 
 test('shiftFlopper', () => {
-  const vec = [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 ]
-  for (const element of shiftFlopper(vec, 3)) {
-    console.log(element)
-  }
+  const vec = init(24, i => i)
+  console.log([ ...shiftFlopper(vec, 1) ])
+  console.log([ ...shiftFlopper(vec, 4) ])
 })